AboutWidget: move layout to .ui file
The widget layout is now in a .ui file.
The controls' contents are still set in the constructor, because these values
will most likely be moved into global constants at a later time and therefore
cannot be set in the Qt Designer.

ImlibWidget is now longer used here.
Qt provides classes and widgets to load and display images, and neither
performance nor file format support is an issue in this case.

A default constructor is added to class KuickUrlWidget, because one is
needed if the class is to be used in .ui files.