cleanup: remove "tagging" of KFileItems
FileWidget::isImage() cannot verify the MIME type of remote files, so
FileWidget::setImage() was used to tag a KFileItem as an image after
downloading and checking the file.

However, KuickShow::showImage() needs this information only during
command line parsing. At that point, it is already known if the file
is an image or not and KuickShow::showImage() can be told through a
parameter to just display the image and ignore the file's MIME type.

Another reason for the change:
The functions KFileItem::setExtraData() and ~::extraData() are no
longer available in KDE5.