port: update make files and code to compile against KDE5
The CMakeLists.txt files are updated to find and use the new KDE5 commands.
The images are renamed due to KDE5's naming policy.

The code is modified only to such extent that it compiles with KDELibs4Support.
But since it relies on it, the porting isn't done yet. The compile process
outputs *a ton* of warnings regarding deprecated classes (most of them KUrl)
which will help in locating legacy code.

Note:
KDE5 apps have a different location where they store their settings. The old
settings are *not* automatically copied to that new location.

Note 2:
Some rough testing has been done.
Browsing directories and viewing images seems to work (plain images and those
located in .zip files, as well as remote images on a web server (http://...)).

Still, until the legacy KDE4 code is completely removed, care should be taken
when using this software.