port: replace KTemporaryFile
KTemporaryFile is part of KDELibs4Support. It is replaced by QTemporaryFile.

Since the convenience functions in KTemporaryFile no longer exist, there's now
a new function in FileCache which does the creation of a QTemporaryFile with
a specific prefix and/or suffix for easier use.
As a side-effect, all temporary files are now located in FileCache's temporary
directory.