port: replace KFileDialog
KFileDialog is part of KDELibs4Support. It is replaced by QFileDialog.

Both instances using KFileDialog used features that are not provided by
QFileDialog and have to be re-implemented:

1) Opening any number of files *and* directories is not possible with plain
   QFileDialog. There's a new class, OpenFilesAndDirsDialog, which provides
   this functionality.

2) When saving an image, the "Save As" dialog had an additional checkbox
   which now has to be manually inserted into QFileDialog.

Because of these changes, native dialogs cannot be used in either case.
The code is also somewhat dependent on the internal structure of QFileDialog
which is bound to cause problems on future Qt releases.