Delayed delete new window if loading an image failed
If it is immediately deleted in KuickShow::showImage() then
a command line loading multiple files, for example

  kuickshow foo.png bar.png

where foo.png does not exist or cannot be loaded will exit the
application as soon as the error box is closed, because the error box
will have entered a nested event loop and there are no application
windows open when it finishes.  The result is that bar.png is never
loaded.