Start to modernise: remove uses of raw Xlib for image window
Use the corresponding Qt facilities instead, including displaying
the image as a QPixmap within a QLabel and a QScrollArea for moving
the image around.  Some changes to the KuickImage rendering and
conversion to accomodate this.

Not sure whether this will immediately allow the application to work
on Wayland, because Imlib 1 still depends on X11 (and requires its
headers at build time).

The brightness/gamma/contrast image modifications are not currently
supported.