Imlib2: Support for image modifications (brightness/contrast/gamma)
More involved than expected, because Imlib2 does not provide any
means to either access the current modifier values (as was simply
available via the ImlibColorModifier struct in Imlib1) or even to
set an absolute value.  The way that the contrast and gamma modifiers
combine cumulatively is also not obvious (it appears to be
multiplicative rather than simple additive as is the case for
brightness).

Move all of the image modifier initialisation and updating into
ImlibWidget, leaving only the GUI slots in ImageWIndow.  Use 3
new "step*Internal" functions to update the values with appropriate
value conversion for each library.  Use these to initialise the values
when loading a new image.  Rename 'mod' to 'myModifier' to make it
easier to search for the name.

Still TODO: the preview images in DefaultsWidget (need a way to
ignore the application config presets and set absolute values).