Manage the Imlib data and application settings in a singleton class
New class ImlibParams with accessors for the Imlib data (only needed
for Imlib 1), the KuickData application settings and the ImData (not
to be confused with ImlibData) image settings.  Do the library
initialisation here instead of both in KuickShow and in ImlibWidget.

Use the global data throughout, instead of passing around and retaining
pointers to the same data everywhere.  Eliminates some member variables
and simplifies constructors of ImageCache, ImlibWidget and ImageWindow.

ImageCache::loadImageWithQt() can now be file-static.

Still to do: DefaultsWidget modifications preview, image saving