fix: restore switch between normal and fullscreen mode
Calling QWidget::move() and QWidget::resize() resets its fullscreen
flag (QWidget::windowState() & Qt::WindowFullScreen) while keeping it
displayed as it is. A subsequent call to QWidget::showNormal() does
nothing since the widget is not (technically) in fullscreen.

Add a check for the image window not being in fullscreen before
changing its position or size.