refactor: remove usages of KDirOperator::actionCollection()
Create a separate QHash property in class KuickShow which contains all
actions created by the class. The action collection of fileWidget is no
longer extended with these.

To configure the shortcuts of both KuickShow's and fileWidget's actions
via the config dialog, a temporary KActionCollection is used while the
dialog is active.